requirements.txt
absl-py==0.12.0
argon2-cffi @ file:///tmp/build/80754af9/argon2-cffi_1613036642480/work
astunparse==1.6.3
async-generator==1.10
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
autoindex==0.3.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
black==21.12b0
bleach @ file:///tmp/build/80754af9/bleach_1628110601003/work
cachetools==4.1.1
certifi==2021.10.8
cffi @ file:///tmp/build/80754af9/cffi_1625814693446/work
chardet==3.0.4
click==7.1.2
cryptography==3.4.7
cycler==0.10.0
decorator @ file:///tmp/build/80754af9/decorator_1632776554403/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
entrypoints==0.3
filelock==3.0.12
Flask==2.0.1
flatbuffers==1.12
gast==0.3.3
gdown==3.12.0
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1638529598726/work
imread==0.7.4
inline==0.0.1
ipykernel @ file:///tmp/build/80754af9/ipykernel_1596206598566/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1635944172793/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
itsdangerous==2.0.1
jedi @ file:///tmp/build/80754af9/jedi_1611333758854/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1635780242639/work
joblib==0.16.0
jsonschema @ file:///Users/ktietz/demo/mc3/conda-bld/jsonschema_1630511932244/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1640335223713/work
jupyter-core @ file:///tmp/build/80754af9/jupyter_core_1636537035214/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1621528142364/work
matplotlib==3.3.0
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mistune @ file:///tmp/build/80754af9/mistune_1594373098390/work
mypy-extensions==0.4.3
nb-conda-kernels==2.2.3
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert @ file:///tmp/build/80754af9/nbconvert_1624479062492/work
nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
networkx==2.6.3
nltk==3.5
notebook @ file:///tmp/build/80754af9/notebook_1596838645154/work
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.4.60
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.1.1
pandocfilters @ file:///tmp/build/80754af9/pandocfilters_1605120451932/work
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
pathspec==0.9.0
pdfminer.six==20201018
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==7.2.0
platformdirs==2.4.1
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1637050397234/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.12.4
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///tmp/build/80754af9/pygments_1629234116488/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
PyPDF2==1.26.0
pyrsistent @ file:///tmp/build/80754af9/pyrsistent_1636098896055/work
PySocks==1.7.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2020.1
PyWavelets==1.2.0
pyzmq @ file:///tmp/build/80754af9/pyzmq_1638434924971/work
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.19.0
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.10.1
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
six @ file:///tmp/build/80754af9/six_1623709665295/work
sortedcontainers==2.4.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
termcolor==1.1.0
terminado==0.9.4
testpath @ file:///tmp/build/80754af9/testpath_1624638946665/work
tf==1.0.0
threadpoolctl==2.1.0
tifffile==2021.11.2
tomli==1.2.3
tornado @ file:///tmp/build/80754af9/tornado_1606942283357/work
tqdm==4.48.2
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typed-ast==1.5.1
typing_extensions==4.0.1
urllib3==1.25.10
vector==0.0.0
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
Werkzeug==2.0.1
wrapt==1.12.1
zipp @ file:///tmp/build/80754af9/zipp_1633618647012/work
